+++ src/entities/run.rs
@@ -2,12 +2,13 @@
 
 use sea_orm::entity::prelude::*;
 
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
+#[derive(Clone, Debug, PartialEq, DeriveEntityModel)]
 #[sea_orm(table_name = "run")]
 pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
-    pub cpu_avg_power: i32,
+    #[sea_orm(column_type = "Float")]
+    pub cpu_avg_power: f32,
     pub start_time: i64,
     pub stop_time: Option<i64>,
 }
